requirements.txt
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
# torchaudio==2.1.2+cu118

# # 指定 PyTorch CUDA 版本的來源
# -f https://download.pytorch.org/whl/torch_stable.html

# 其餘套件
jax
datasets
accelerate==0.30.1
deepspeed==0.13.1
Levenshtein
matplotlib
peft==0.14.0
pillow
jupyterlab
opencv-python
ultralytics
tensorboard
pycocotools
scikit-image
bitsandbytes==0.43.1
ninja
selenium==4.12.0
transformers==4.49.0
qwen-vl-utils
scikit-learn
nltk
wandb
openai
numpy<2.0